SurveyRefData: moved to refData folder
