+++ src/survey/refData/refData.ts
@@ -1,0 +1,9 @@
+import { CategoryItem } from '../../category'
+import { Taxon } from '../../taxonomy'
+
+export interface SurveyRefData {
+  categoryItemUuidIndex: { [categoryUuid: string]: { [parentItemUuid: string]: { [code: string]: string } } }
+  categoryItemIndex: { [categoryItemUuid: string]: CategoryItem }
+  taxonUuidIndex: { [taxonomyUuid: string]: { [taxonCode: string]: Taxon } }
+  taxonIndex: { [taxonUuid: string]: Taxon }
+}
